Removing relative permalinks for Jekyll
The page build failed with the following error:

Your site is using the `relative_permalinks` configuration option. This setting is deprecated as has been removed from the latest version of Jekyll. To ensure your site continues to build as expected, remove the option from your site's configuration and update any post or page permalinks to be absolute to the site root, not the parent folder. For more information, see https://help.github.com/articles/removing-relative-permalinks.

GitHub Pages was recently upgraded to Jekyll 3.0. It may help to confirm you're using the correct dependencies:

 https://github.com/blog/2100-github-pages-now-faster-and-simpler-with-jekyll-3-0